requirements.txt
torch
torchvision
opencv-python
Pillow
matplotlib
seaborn
scikit-learn
reportlab